refactor: move picker popup
